intimg: fixes from comments
