+++ ttnn/cpp/ttnn/operations/experimental/reduction/integral_image/device/kernels/intimg_compute.cpp
@@ -20,103 +20,9 @@
 #include "compute_kernel_api/common.h"
 #include "compute_kernel_api/eltwise_binary_sfpu.h"
 
+#include "common.hpp"
+
 namespace {
-
-constexpr uint32_t ONE_TILE{1};
-constexpr uint32_t FIRST_TILE{0};
-constexpr uint32_t WORKING_REG{0};
-
-/**
- * @brief RAII guard for safely reading from a Circular Buffer (CB).
- *
- * `ReadCBGuard` automatically manages CB read synchronization in a scoped manner.
- * When constructed, it waits for `tiles` tiles to be available at the CB front
- * (`cb_wait_front`), ensuring data readiness before access. Upon destruction,
- * it automatically pops those tiles from the CB front (`cb_pop_front`),
- * signaling that the tiles have been consumed.
- *
- * This guarantees balanced `cb_wait_front`/`cb_pop_front` calls, even in the
- * presence of early returns or exceptions, preventing CB underflows and race
- * conditions.
- *
- * @note This class is strictly non-copyable and non-movable to prevent any
- *       double-pop or premature release of CB resources.
- *
- * Example usage:
- * @code
- * {
- *     ReadCBGuard guard(cb_id, num_tiles);
- *     // Safely read from CB: data is guaranteed ready.
- *     // ...
- * } // Automatically pops the tiles on scope exit.
- * @endcode
- */
-class ReadCBGuard {
-    uint32_t cb;
-    uint32_t tiles;
-
-public:
-    ReadCBGuard(uint32_t cb, uint32_t tiles) : cb{cb}, tiles{tiles} { cb_wait_front(cb, tiles); }
-    ~ReadCBGuard() { cb_pop_front(cb, tiles); }
-
-    ReadCBGuard(const ReadCBGuard&) = delete;  // can not allow to touch the object anyhow
-    ReadCBGuard(ReadCBGuard&&) = delete;
-    ReadCBGuard& operator=(const ReadCBGuard&) = delete;
-    ReadCBGuard& operator=(ReadCBGuard&&) = delete;
-    template <typename any_type_t>
-    operator any_type_t() = delete;
-};
-
-/**
- * @brief RAII guard for safely writing to a Circular Buffer (CB).
- *
- * `WriteCBGuard` automatically manages CB write synchronization in a scoped manner.
- * When constructed, it reserves space for `tiles` tiles at the CB back
- * (`cb_reserve_back`), ensuring sufficient room for writing. Upon destruction,
- * it automatically pushes those tiles to the CB back (`cb_push_back`),
- * making them visible to downstream consumers.
- *
- * This ensures balanced `cb_reserve_back`/`cb_push_back` calls and prevents
- * buffer overflows or mismatched producer-consumer behavior.
- *
- * @note Like `ReadCBGuard`, this class is non-copyable and non-movable to ensure
- *       one-to-one ownership of the CB reservation.
- *
- * Example usage:
- * @code
- * {
- *     WriteCBGuard guard(cb_id, num_tiles);
- *     // Safely write to CB: space is guaranteed reserved.
- *     // ...
- * } // Automatically pushes tiles to the CB on scope exit.
- * @endcode
- */
-class WriteCBGuard {
-    uint32_t cb;
-    uint32_t tiles;
-
-public:
-    WriteCBGuard(uint32_t cb, uint32_t tiles) : cb{cb}, tiles{tiles} { cb_reserve_back(cb, tiles); }
-    ~WriteCBGuard() { cb_push_back(cb, tiles); }
-
-    WriteCBGuard(const WriteCBGuard&) = delete;  // can not allow to touch the object anyhow
-    WriteCBGuard(WriteCBGuard&&) = delete;
-    WriteCBGuard& operator=(const WriteCBGuard&) = delete;
-    WriteCBGuard& operator=(WriteCBGuard&&) = delete;
-    template <typename any_type_t>
-    operator any_type_t() = delete;
-};
-
-FORCE_INLINE uint32_t get_coord_from_tile_xy(uint32_t read_i, uint32_t write_i) {
-    return ((write_i & 0x10) << 5)    // y_hi * 512
-           | ((read_i & 0x10) << 4)   // x_hi * 256
-           | ((write_i & 0x0F) << 4)  // y_lo * 16
-           | (read_i & 0x0F);
-}
-
-FORCE_INLINE constexpr uint32_t block_depth_ceil(uint32_t value, uint32_t block_depth = 32) {
-    return (value + block_depth - 1) / block_depth;
-}
 
 struct IntImgComputeCTAs {
     const uint32_t start_cb;
